+++ src/signed_distance_field/shapes/superellipsoid.rs
@@ -1,4 +1,4 @@
-use rust_gpu_bridge::prelude::Vec3;
+use rust_gpu_bridge::prelude::{Vec3, Abs, Pow};
 use type_fields::Field;
 
 use crate::signed_distance_field::{attributes::distance::Distance, SignedDistanceField};
@@ -17,8 +17,8 @@ impl Default for Superellipsoid {
 
 impl SignedDistanceField<Vec3, Distance> for Superellipsoid {
     fn evaluate(&self, p: Vec3) -> Distance {
-        let d = (p.x.abs().powf(self.e1) + p.y.abs().powf(self.e2)).powf(self.e2 / self.e1)
-            + p.z.abs().powf(self.e1);
+        let d = (p.x.abs().pow(self.e1) + p.y.abs().pow(self.e2)).pow(self.e2 / self.e1)
+            + p.z.abs().pow(self.e1);
         let d = d - 1.0;
         Distance(d)
     }
